requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
qrcode==7.4.2
pillow==10.2.0
reportlab==4.1.0
gunicorn==21.2.0